pyproject.toml
[tool.poetry]
name = "bcraftstatistic"
version = "0.1.0"
description = ""
authors = ["Ra1ze505 <87398772+Ra1ze505@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.9"
djangorestframework = "3.12.4"
django-filter = "21.1"
drf-yasg = "1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
